[Calc] move DimensionOfHierarchy to calc
because is is used multiple times
