+++ mondrian/src/main/java/org/eclipse/daanse/olap/calc/base/type/dimension/DimensionOfHierarchyCalc.java
@@ -11,7 +11,7 @@
 *   SmartCity Jena - initial
 *   Stefan Bischof (bipolis.org) - initial
 */
-package org.eclipse.daanse.olap.function.def.dimension.hierarchy;
+package org.eclipse.daanse.olap.calc.base.type.dimension;
 
 import org.eclipse.daanse.olap.api.Evaluator;
 import org.eclipse.daanse.olap.api.element.Dimension;
